# server/swagger.json
{
    "openapi": "3.0.4",
    "info": {
        "title": "Quiz API",
        "description": "API documentation generated from src/modules controllers and DTOs (test-management, question-management, test-execution and identity-access).",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8031/api",
            "description": "Main api server"
        }
    ],
...
    "paths": {
...
        "/question/{testId}/questions": {
            "post": {
                "tags": ["question"],
                "summary": "Create question",
                "security": [{ "BearerAuth": [] }],
                "parameters": [
                    {
                        "name": "testId",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": { "$ref": "#/components/schemas/QuestionCreateRequestBody" }
                        },
                        "multipart/form-data": {
                            "schema": { "$ref": "#/components/schemas/QuestionCreateMultipartRequestBody" }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Question created",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/QuestionResponse" }
                            }
                        }
                    },
                    "401": { "description": "Unauthorized" },
                    "403": { "description": "Forbidden" },
                    "404": { "description": "Not found" },
                    "422": { "description": "Validation failed, invalid image type, or image too large" }
                }
            }
        },
        "/question/{testId}/questions/{questionId}": {
            "patch": {
                "tags": ["question"],
                "summary": "Update question",
                "security": [{ "BearerAuth": [] }],
                "parameters": [
                    {
                        "name": "testId",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    },
                    {
                        "name": "questionId",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": { "$ref": "#/components/schemas/QuestionUpdateRequestBody" }
                        },
                        "multipart/form-data": {
                            "schema": { "$ref": "#/components/schemas/QuestionUpdateMultipartRequestBody" }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Question updated",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/QuestionResponse" }
                            }
                        }
                    },
                    "401": { "description": "Unauthorized" },
                    "403": { "description": "Forbidden" },
                    "404": { "description": "Not found" },
                    "422": { "description": "Validation failed, invalid image type, or image too large" }
                }
            },
            "delete": {
                "tags": ["question"],
                "summary": "Delete question",
                "security": [{ "BearerAuth": [] }],
                "parameters": [
                    {
                        "name": "testId",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    },
                    {
                        "name": "questionId",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "responses": {
                    "204": { "description": "No content" },
                    "401": { "description": "Unauthorized" },
                    "403": { "description": "Forbidden" },
                    "404": { "description": "Not found" }
                }
            }
        },
        "/question/{testId}/questions/{questionId}/order": {
            "patch": {
                "tags": ["question"],
                "summary": "Change question order",
                "security": [{ "BearerAuth": [] }],
                "parameters": [
                    {
                        "name": "testId",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    },
                    {
                        "name": "questionId",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": { "$ref": "#/components/schemas/QuestionChangeOrderRequestBody" }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Question order changed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": { "$ref": "#/components/schemas/QuestionResponse" }
                                }
                            }
                        }
                    },
                    "401": { "description": "Unauthorized" },
                    "403": { "description": "Forbidden" },
                    "404": { "description": "Not found" },
                    "422": { "description": "Validation failed" }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
        "schemas": {
...
            "QuestionConfigOption": {
                "type": "object",
                "required": ["id", "value"],
                "properties": {
                    "id": { "type": "string" },
                    "value": { "type": "string" }
                }
            },
            "QuestionConfigOptionArray": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/QuestionConfigOption" }
            },
            "QuestionConfigInput": {
                "type": "object",
                "required": ["answer", "ignore_case", "type"],
                "properties": {
                    "type": { "type": "string", "enum": ["input"] },
                    "answer": { "type": "string" },
                    "ignore_case": { "type": "boolean" }
                }
            },
            "QuestionConfigSingleChoice": {
                "type": "object",
                "required": ["answer", "options", "type"],
                "properties": {
                    "type": { "type": "string", "enum": ["single_choice"] },
                    "answer": { "type": "string" },
                    "options": { "$ref": "#/components/schemas/QuestionConfigOptionArray" }
                }
            },
            "QuestionConfigMultipleChoice": {
                "type": "object",
                "required": ["answer", "options", "type"],
                "properties": {
                    "type": { "type": "string", "enum": ["multiple_choice"] },
                    "answer": {
                        "type": "array",
                        "items": { "type": "string" }
                    },
                    "options": { "$ref": "#/components/schemas/QuestionConfigOptionArray" }
                }
            },
            "QuestionConfigOrderAnswerOption": {
                "type": "object",
                "required": ["optionId", "order"],
                "properties": {
                    "optionId": { "type": "string" },
                    "order": { "type": "number" }
                }
            },
            "QuestionConfigOrder": {
                "type": "object",
                "required": ["answer", "options", "type"],
                "properties": {
                    "type": { "type": "string", "enum": ["order"] },
                    "answer": {
                        "type": "array",
                        "items": { "$ref": "#/components/schemas/QuestionConfigOrderAnswerOption" }
                    },
                    "options": { "$ref": "#/components/schemas/QuestionConfigOptionArray" }
                }
            },
            "QuestionRequestConfig": {
                "oneOf": [
                    { "$ref": "#/components/schemas/QuestionConfigInput" },
                    { "$ref": "#/components/schemas/QuestionConfigSingleChoice" },
                    { "$ref": "#/components/schemas/QuestionConfigMultipleChoice" },
                    { "$ref": "#/components/schemas/QuestionConfigOrder" }
                ],
                "discriminator": { "propertyName": "type" }
            },
...
            "QuestionCreateRequestBody": {
                "type": "object",
                "required": ["description", "config"],
                "properties": {
                    "description": { "type": "string" },
                    "config": { "$ref": "#/components/schemas/QuestionRequestConfig" },
                    "score": {
                        "type": "number",
                        "description": "Points awarded for a correct answer. Defaults to 1."
                    }
                }
            },
            "QuestionCreateMultipartRequestBody": {
                "type": "object",
                "required": ["description", "config"],
                "properties": {
                    "description": { "type": "string" },
                    "config": {
                        "type": "string",
                        "description": "JSON string with question config"
                    },
                    "score": {
                        "type": "number",
                        "description": "Points awarded for a correct answer. Defaults to 1."
                    },
                    "image": {
                        "type": "string",
                        "format": "binary",
                        "description": "Optional image file (max 5MB, image/* MIME type)"
                    }
                }
            },
            "QuestionUpdateRequestBody": {
                "type": "object",
                "required": ["description", "config"],
                "properties": {
                    "description": { "type": "string" },
                    "config": { "$ref": "#/components/schemas/QuestionRequestConfig" },
                    "score": {
                        "type": "number",
                        "description": "Points awarded for a correct answer. Defaults to 1."
                    }
                }
            },
            "QuestionUpdateMultipartRequestBody": {
                "type": "object",
                "required": ["description", "config"],
                "properties": {
                    "description": { "type": "string" },
                    "config": {
                        "type": "string",
                        "description": "JSON string with question config"
                    },
                    "score": {
                        "type": "number",
                        "description": "Points awarded for a correct answer. Defaults to 1."
                    },
                    "image": {
                        "type": "string",
                        "format": "binary",
                        "description": "Optional image file (max 5MB, image/* MIME type)"
                    }
                }
            },
            "QuestionChangeOrderRequestBody": {
                "type": "object",
                "properties": {
                    "previous_question_id": {
                        "type": "string",
                        "nullable": true,
                        "description": "ID of the question that should appear immediately before the moved question"
                    },
                    "next_question_id": {
                        "type": "string",
                        "nullable": true,
                        "description": "ID of the question that should appear immediately after the moved question"
                    }
                }
            },
            "QuestionResponse": {
                "type": "object",
                "required": ["id", "test_id", "sort_key", "description", "score", "image", "config"],
                "properties": {
                    "id": { "type": "string" },
                    "test_id": { "type": "string" },
                    "sort_key": { "type": "number" },
                    "description": { "type": "string" },
                    "score": {
                        "type": "number",
                        "description": "Points awarded for a correct answer"
                    },
                    "image": {
                        "type": "string",
                        "nullable": true,
                        "description": "Relative URL to the question image, or null"
                    },
                    "config": { "$ref": "#/components/schemas/QuestionRequestConfig" }
                }
            },
...
        }
    }
}